chore: remove blog list page
