+++ apps/web/vibes/soul/examples/pages/blog/index.tsx
@@ -1,5 +1,129 @@
-import { BlogPage } from '@/vibes/soul/pages/blog'
+import { posts } from '@/vibes/soul/examples/sections/blog-post-list'
+import { headerLinks } from '@/vibes/soul/examples/sections/header/electric'
+import { AnnouncementBar } from '@/vibes/soul/primitives/announcement-bar'
+import { FeaturedBlogPostList } from '@/vibes/soul/sections/featured-blog-post-list'
+import { Footer } from '@/vibes/soul/sections/footer'
+import {
+  Amex,
+  ApplePay,
+  Bitcoin,
+  GooglePay,
+  Mastercard,
+  Paypal,
+  Visa,
+} from '@/vibes/soul/sections/footer/payment-icons'
+import { Facebook, Instagram, X, Youtube } from '@/vibes/soul/sections/footer/social-icons'
+import { Header } from '@/vibes/soul/sections/header'
+import { Subscribe } from '@/vibes/soul/sections/subscribe'
+
+const socialMediaLinks = [
+  {
+    href: '#',
+    icon: <Facebook />,
+  },
+  {
+    href: '#',
+    icon: <X />,
+  },
+  {
+    href: '#',
+    icon: <Instagram />,
+  },
+  {
+    href: '#',
+    icon: <Youtube />,
+  },
+]
+
+const paymentIconsArray: React.ReactNode[] = [
+  <Visa key="Visa" />,
+  <Amex key="Amex" />,
+  <Mastercard key="Mastercard" />,
+  <Paypal key="Paypal" />,
+  <GooglePay key="GooglePay" />,
+  <ApplePay key="ApplePay" />,
+  <Bitcoin key="Bitcoin" />,
+]
+
+const locales = [
+  { id: '1', region: 'US', language: 'EN' },
+  { id: '2', region: 'FR', language: 'FR' },
+  { id: '3', region: 'DE', language: 'DC' },
+  { id: '4', region: 'IT', language: 'IT' },
+]
 
 export default function Preview() {
-  return <BlogPage />
+  return (
+    <>
+      <AnnouncementBar>
+        Get <strong>15% off</strong> and free shipping with discount code{' '}
+        <strong>&quot;welcome&quot;</strong>
+      </AnnouncementBar>
+
+      <Header
+        links={headerLinks}
+        logo="SOUL"
+        cartHref="#"
+        accountHref="#"
+        locales={locales}
+        activeLocale="EN"
+      />
+
+      <FeaturedBlogPostList
+        title="Plant Life"
+        description="Expert Tips & Inspiration for Every Plant Lover"
+        cta={{ href: '#', label: 'View All' }}
+        posts={posts}
+      />
+
+      <Subscribe
+        image={{
+          src: 'https://rstr.in/monogram/vibes/m12FEyfnuDl',
+          alt: 'Lady with plant',
+        }}
+        title="Sign up for our newsletter"
+        description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor."
+      />
+
+      <Footer
+        logo="SOUL"
+        sections={[
+          {
+            title: 'Categories',
+            links: [
+              { label: 'Coats & Jackets', href: '#' },
+              { label: 'T-Shirts', href: '#' },
+              { label: 'Sweatshirts', href: '#' },
+              { label: 'Pants', href: '#' },
+            ],
+          },
+          {
+            title: 'Company',
+            links: [
+              { label: 'About', href: '#' },
+              { label: 'Stories', href: '#' },
+              { label: 'Careers', href: '#' },
+              { label: 'Stores', href: '#' },
+            ],
+          },
+          {
+            title: 'Help & Support',
+            links: [
+              { label: 'FAQs', href: '#' },
+              { label: 'Contact Us', href: '#' },
+              { label: 'Returns', href: '#' },
+              { label: 'Shipping', href: '#' },
+            ],
+          },
+        ]}
+        contactInformation={{
+          address: 'info@mywebsite.com',
+          phone: '+(1)408 123 4567',
+        }}
+        copyright={`© ${new Date().getFullYear().toString()} SOUL - Powered by Monogram`}
+        paymentIcons={paymentIconsArray}
+        socialMediaLinks={socialMediaLinks}
+      />
+    </>
+  )
 }
